Refactor + Css + text area ajustments
Added css to make the app look a bit retro might add more style later.
Added some basic padding in all the views to space things out.
Added new rules to the text area so now it goes to a new row instead of writing forever in one.
Cleaned the code a bit a lot of  empty new lines.